grp12_ueb04/test/ueb04/WaypointTest.java
package ueb04;

import org.junit.Test;
import static org.junit.Assert.*;

/**
 *
 * @author GeritKaleck
 */
public class WaypointTest {
    
    //<editor-fold defaultstate="collapsed" desc="Test of constructors">
    /**
     * Test of constructors, of class Waypoint.
     */
    @Test
    public void testConstructor() {
        // the default-constructor is just expected to exist.
        Waypoint waypoint = new Waypoint();
        assertEquals(0, waypoint.getX());
        assertEquals(0, waypoint.getY());
    }
    
    @Test
    public void testConstructor_int_int() {
        Waypoint waypoint = new Waypoint(1, 2);
        assertEquals(1, waypoint.getX());
        assertEquals(2, waypoint.getY());
    }
//</editor-fold>

    
}
